pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "scripts"
version = "0.1.0"
description = "Protocols and scripts for the lab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.0",
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
deepseek-chat = "scripts.deepseek_chat:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai", "src/scripts"]

[tool.uv]
# uv (fast package manager, optional)

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = ["black>=25.9.0", "mypy>=1.18.2", "ruff>=0.14.0"]
